+++ internal/app/balance/models/models.go
@@ -7,11 +7,11 @@ const (
 
 type Balance struct {
 	ID     int64
-	Amount int64
+	Amount float64
 }
 type TransferMoney struct {
 	SenderID        int64
-	SenderBalance   int64
+	SenderBalance   float64
 	ReceiverID      int64
-	ReceiverBalance int64
+	ReceiverBalance float64
 }
